Add ability to rotate images based on EXIF data
Look at me using Yarn and stuff.

It was about time to extract the JavaScript code to its own file.

The only simple way I found so far to rotate images appropriately was to use
classes and just have some CSS that knows how to rotate. The nice thing is that
it almost work for both thumbnails and large images without much work.