Dramatically speed up EXIF data loading
Instead of loading the file as a DataURL to create the image and then
converting it to an ArrayBuffer to read the EXIF data I’m directly asking for
an ArrayBuffer from the FileReader which speeds up the EXIF reading.

I’m also removing the event listeners for FileReader’s load event to prevent
bloat.